requirements.txt
fastapi
uvicorn
sqlalchemy
bcrypt
cryptography
python-jose
python-multipart
passlib
psycopg2